factorize tool call view
